remove old update handler
