+++ nuget/helpers/lib/NuGetUpdater/NuGetUpdater.Core.Test/Run/EndToEndTests.cs
@@ -1,20 +1,26 @@
+using System.Collections.Immutable;
 using System.Text;
+using System.Text.Json;
 
-using NuGetUpdater.Core.Run.ApiModel;
-using NuGetUpdater.Core.Run;
-using Xunit;
 using NuGetUpdater.Core.Analyze;
 using NuGetUpdater.Core.Discover;
-using System.Collections.Immutable;
+using NuGetUpdater.Core.Run;
+using NuGetUpdater.Core.Run.ApiModel;
+using NuGetUpdater.Core.Test.Update;
+
+using Xunit;
 
 namespace NuGetUpdater.Core.Test.Run;
+
+using TestFile = (string Path, string Content);
+using RawTestFile = (string Path, byte[] Content);
 
 public class EndToEndTests
 {
     [Fact]
     public async Task WithPackageReference()
     {
-        await RunWorkerTests.RunAsync(
+        await RunAsync(
             experimentsManager: new ExperimentsManager() { UseDirectDiscovery = true },
             packages: [
                 MockNuGetPackage.CreateSimplePackage("Some.Package", "1.0.0", "net9.0"),
@@ -53,11 +59,6 @@ public class EndToEndTests
             discoveryWorker: null, // use real worker
             analyzeWorker: null, // use real worker
             updaterWorker: null, // use real worker
-            expectedResult: new()
-            {
-                Base64DependencyFiles = [],
-                BaseCommitSha = "TEST-COMMIT-SHA",
-            },
             expectedApiMessages: [
                 new IncrementMetric()
                 {
@@ -140,9 +141,9 @@ public class EndToEndTests
                         },
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = TestPullRequestCommitMessage,
+                    PrTitle = TestPullRequestTitle,
+                    PrBody = TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA")
@@ -153,7 +154,7 @@ public class EndToEndTests
     [Fact]
     public async Task WithPackagesConfig()
     {
-        await RunWorkerTests.RunAsync(
+        await RunAsync(
             experimentsManager: new ExperimentsManager() { UseDirectDiscovery = true },
             packages: [
                 MockNuGetPackage.CreateSimplePackage("Some.Package", "1.0.0", "net45"),
@@ -207,11 +208,6 @@ public class EndToEndTests
             discoveryWorker: null, // use real worker
             analyzeWorker: null, // use real worker
             updaterWorker: null, // use real worker
-            expectedResult: new()
-            {
-                Base64DependencyFiles = [],
-                BaseCommitSha = "TEST-COMMIT-SHA",
-            },
             expectedApiMessages: [
                 new IncrementMetric()
                 {
@@ -311,9 +307,9 @@ public class EndToEndTests
                         },
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = TestPullRequestCommitMessage,
+                    PrTitle = TestPullRequestTitle,
+                    PrBody = TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA")
@@ -325,7 +321,7 @@ public class EndToEndTests
     public async Task LegacyProject_With_PackageReference()
     {
         var experimentsManager = new ExperimentsManager() { UseDirectDiscovery = true };
-        await RunWorkerTests.RunAsync(
+        await RunAsync(
             experimentsManager: experimentsManager,
             packages: [
                 MockNuGetPackage.CreateSimplePackage("Some.Package", "1.0.0", "net45"),
@@ -393,11 +389,6 @@ public class EndToEndTests
             }),
             analyzeWorker: null, // use real worker
             updaterWorker: null, // use real worker
-            expectedResult: new()
-            {
-                Base64DependencyFiles = [],
-                BaseCommitSha = "TEST-COMMIT-SHA",
-            },
             expectedApiMessages: [
                 new IncrementMetric()
                 {
@@ -483,9 +474,9 @@ public class EndToEndTests
                         },
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = TestPullRequestCommitMessage,
+                    PrTitle = TestPullRequestTitle,
+                    PrBody = TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA")
@@ -493,10 +484,8 @@ public class EndToEndTests
         );
     }
 
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task UpdatePackageWithDifferentVersionsInDifferentDirectories(bool useLegacyUpdateHandler)
+    [Fact]
+    public async Task UpdatePackageWithDifferentVersionsInDifferentDirectories()
     {
         // this test passes `null` for discovery, analyze, and update workers to fully test the desired behavior
 
@@ -504,88 +493,8 @@ public class EndToEndTests
         //   library1.csproj - top level dependency, already up to date
         //   library2.csproj - top level dependency, needs direct update
         //   library3.csproj - transitive dependency, needs pin
-        var base64DependencyFiles = useLegacyUpdateHandler
-            ? new[]
-            {
-                new DependencyFile()
-                    {
-                        Directory = "/",
-                        Name = "Directory.Build.props",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("<Project />")),
-                        ContentEncoding = "base64",
-                    },
-                    new DependencyFile()
-                    {
-                        Directory = "/",
-                        Name = "Directory.Build.targets",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("<Project />")),
-                        ContentEncoding = "base64",
-                    },
-                    new DependencyFile()
-                    {
-                        Directory = "/",
-                        Name = "Directory.Packages.props",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("""
-                            <Project>
-                              <PropertyGroup>
-                                <ManagePackageVersionsCentrally>false</ManagePackageVersionsCentrally>
-                              </PropertyGroup>
-                            </Project>
-                            """)),
-                        ContentEncoding = "base64",
-                    },
-                    new DependencyFile()
-                    {
-                        Directory = "/library1",
-                        Name = "library1.csproj",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("""
-                            <Project Sdk="Microsoft.NET.Sdk">
-                              <PropertyGroup>
-                                <TargetFramework>net8.0</TargetFramework>
-                              </PropertyGroup>
-                              <ItemGroup>
-                                <PackageReference Include="Some.Package" Version="2.0.0" />
-                              </ItemGroup>
-                            </Project>
-                            """)),
-                        ContentEncoding = "base64",
-                    },
-                    new DependencyFile()
-                    {
-                        Directory = "/library2",
-                        Name = "library2.csproj",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("""
-                            <Project Sdk="Microsoft.NET.Sdk">
-                              <PropertyGroup>
-                                <TargetFramework>net8.0</TargetFramework>
-                              </PropertyGroup>
-                              <ItemGroup>
-                                <PackageReference Include="Some.Package" Version="1.0.0" />
-                              </ItemGroup>
-                            </Project>
-                            """)),
-                        ContentEncoding = "base64",
-                    },
-                    new DependencyFile()
-                    {
-                        Directory = "/library3",
-                        Name = "library3.csproj",
-                        Content = Convert.ToBase64String(Encoding.UTF8.GetBytes("""
-                            <Project Sdk="Microsoft.NET.Sdk">
-                              <PropertyGroup>
-                                <TargetFramework>net8.0</TargetFramework>
-                              </PropertyGroup>
-                              <ItemGroup>
-                                <PackageReference Include="Package.With.Transitive.Dependency" Version="0.1.0" />
-                              </ItemGroup>
-                            </Project>
-                            """)),
-                        ContentEncoding = "base64",
-                    }
-            }
-            : [];
-        await RunWorkerTests.RunAsync(
-            experimentsManager: new ExperimentsManager() { UseDirectDiscovery = true, UseLegacyUpdateHandler = useLegacyUpdateHandler },
+        await RunAsync(
+            experimentsManager: new ExperimentsManager() { UseDirectDiscovery = true },
             packages: [
                 MockNuGetPackage.CreateSimplePackage("Some.Package", "1.0.0", "net8.0"),
                 MockNuGetPackage.CreateSimplePackage("Some.Package", "2.0.0", "net8.0"),
@@ -666,11 +575,6 @@ public class EndToEndTests
             discoveryWorker: null,
             analyzeWorker: null,
             updaterWorker: null,
-            expectedResult: new RunResult()
-            {
-                Base64DependencyFiles = base64DependencyFiles,
-                BaseCommitSha = "TEST-COMMIT-SHA",
-            },
             expectedApiMessages: [
                 new UpdatedDependencyList()
                 {
@@ -837,13 +741,112 @@ public class EndToEndTests
                         }
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = TestPullRequestCommitMessage,
+                    PrTitle = TestPullRequestTitle,
+                    PrBody = TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA")
             ]
         );
+    }
+
+    public const string TestPullRequestCommitMessage = "test-pull-request-commit-message";
+    public const string TestPullRequestTitle = "test-pull-request-title";
+    public const string TestPullRequestBody = "test-pull-request-body";
+
+    internal static Task RunAsync(Job job, TestFile[] files, IDiscoveryWorker? discoveryWorker, IAnalyzeWorker? analyzeWorker, IUpdaterWorker? updaterWorker, object[] expectedApiMessages, MockNuGetPackage[]? packages = null, ExperimentsManager? experimentsManager = null, string? repoContentsPath = null)
+    {
+        var rawTestFiles = files.Select(f => (f.Path, Encoding.UTF8.GetBytes(f.Content))).ToArray();
+        return RunAsync(job, rawTestFiles, discoveryWorker, analyzeWorker, updaterWorker, expectedApiMessages, packages, experimentsManager, repoContentsPath);
+    }
+
+    private static async Task RunAsync(Job job, RawTestFile[] rawFiles, IDiscoveryWorker? discoveryWorker, IAnalyzeWorker? analyzeWorker, IUpdaterWorker? updaterWorker, object[] expectedApiMessages, MockNuGetPackage[]? packages = null, ExperimentsManager? experimentsManager = null, string? repoContentsPath = null)
+    {
+        // arrange
+        using var tempDirectory = new TemporaryDirectory();
+        repoContentsPath ??= tempDirectory.DirectoryPath;
+        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory(packages, repoContentsPath);
+        foreach (var (path, content) in rawFiles)
+        {
+            var fullPath = Path.Combine(repoContentsPath, path);
+            var directory = Path.GetDirectoryName(fullPath)!;
+            Directory.CreateDirectory(directory);
+            await File.WriteAllBytesAsync(fullPath, content);
+        }
+
+        // act
+        experimentsManager ??= new ExperimentsManager();
+        var testApiHandler = new TestApiHandler();
+        var logger = new TestLogger();
+        var jobId = "TEST-JOB-ID";
+        discoveryWorker ??= new DiscoveryWorker(jobId, experimentsManager, logger);
+        analyzeWorker ??= new AnalyzeWorker(jobId, experimentsManager, logger);
+        updaterWorker ??= new UpdaterWorker(jobId, experimentsManager, logger);
+
+        var worker = new RunWorker(jobId, testApiHandler, discoveryWorker, analyzeWorker, updaterWorker, logger);
+        var repoContentsPathDirectoryInfo = new DirectoryInfo(repoContentsPath);
+        await worker.RunAsync(job, repoContentsPathDirectoryInfo, repoContentsPathDirectoryInfo, "TEST-COMMIT-SHA", experimentsManager);
+        var actualApiMessages = testApiHandler.ReceivedMessages
+            .Select(m =>
+            {
+                object newObject;
+                switch (m.Object)
+                {
+                    // this isn't the place to verify the generated text
+                    case CreatePullRequest create:
+                        newObject = create with { CommitMessage = TestPullRequestCommitMessage, PrTitle = TestPullRequestTitle, PrBody = TestPullRequestBody };
+                        break;
+                    case UpdatePullRequest update:
+                        newObject = update with { CommitMessage = TestPullRequestCommitMessage, PrTitle = TestPullRequestTitle, PrBody = TestPullRequestBody };
+                        break;
+                    // don't test callstacks
+                    case UnknownError unknown:
+                        var message = (string)unknown.Details["error-message"];
+                        var stackTraceOffset = message.IndexOf('\n');
+                        if (stackTraceOffset >= 0)
+                        {
+                            var messageWithoutStackTrace = message[..stackTraceOffset].TrimEnd('\r');
+                            unknown.Details["error-message"] = messageWithoutStackTrace;
+                        }
+
+                        newObject = unknown;
+                        break;
+                    default:
+                        newObject = m.Object;
+                        break;
+                }
+
+                return (m.Type, Object: newObject);
+            }
+            ).ToArray();
+
+        // assert
+        for (int i = 0; i < Math.Min(actualApiMessages.Length, expectedApiMessages.Length); i++)
+        {
+            var actualMessage = actualApiMessages[i];
+            var expectedMessage = expectedApiMessages[i];
+            Assert.Equal(expectedMessage.GetType(), actualMessage.Type);
+
+            var expectedContent = SerializeObjectAndType(expectedMessage);
+            var actualContent = SerializeObjectAndType(actualMessage.Object);
+            Assert.Equal(expectedContent, actualContent);
+        }
+
+        if (actualApiMessages.Length > expectedApiMessages.Length)
+        {
+            var extraApiMessages = actualApiMessages.Skip(expectedApiMessages.Length).Select(m => SerializeObjectAndType(m.Object)).ToArray();
+            Assert.Fail($"Expected {expectedApiMessages.Length} API messages, but got {extraApiMessages.Length} extra:\n\t{string.Join("\n\t", extraApiMessages)}");
+        }
+        if (expectedApiMessages.Length > actualApiMessages.Length)
+        {
+            var missingApiMessages = expectedApiMessages.Skip(actualApiMessages.Length).Select(m => SerializeObjectAndType(m)).ToArray();
+            Assert.Fail($"Expected {expectedApiMessages.Length} API messages, but only got {actualApiMessages.Length}; missing:\n\t{string.Join("\n\t", missingApiMessages)}");
+        }
+    }
+
+    internal static string SerializeObjectAndType(object obj)
+    {
+        return $"{obj.GetType().Name}:{JsonSerializer.Serialize(obj, RunWorker.SerializerOptions)}";
     }
 }
