+++ nuget/helpers/lib/NuGetUpdater/NuGetUpdater.Core.Test/Run/UpdateHandlers/CreateSecurityUpdatePullRequestHandlerTests.cs
@@ -136,9 +136,9 @@ public class CreateSecurityUpdatePullRequestHandlerTests : UpdateHandlersTestsBa
                         }
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = EndToEndTests.TestPullRequestCommitMessage,
+                    PrTitle = EndToEndTests.TestPullRequestTitle,
+                    PrBody = EndToEndTests.TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA"),
@@ -283,9 +283,9 @@ public class CreateSecurityUpdatePullRequestHandlerTests : UpdateHandlersTestsBa
                         }
                     ],
                     BaseCommitSha = "TEST-COMMIT-SHA",
-                    CommitMessage = RunWorkerTests.TestPullRequestCommitMessage,
-                    PrTitle = RunWorkerTests.TestPullRequestTitle,
-                    PrBody = RunWorkerTests.TestPullRequestBody,
+                    CommitMessage = EndToEndTests.TestPullRequestCommitMessage,
+                    PrTitle = EndToEndTests.TestPullRequestTitle,
+                    PrBody = EndToEndTests.TestPullRequestBody,
                     DependencyGroup = null,
                 },
                 new MarkAsProcessed("TEST-COMMIT-SHA"),
